+++ SecretSharing/SecretSharing/UserRating.cs
@@ -4,6 +4,6 @@
     {
         public int UserId { get; set; }
         public int ItemId { get; set; }
-        public int Rating { get; set; }
+        public sbyte Rating { get; set; }
     }
 }
